Changes & improvements in additional certificates code
