--- src/common/types/proxyFields.ts
@@ -1,8 +1,0 @@
-import { Cluster } from '../api';
-
-export type ProxyFieldsType = {
-  enableProxy: boolean;
-  httpProxy?: Cluster['httpProxy'];
-  httpsProxy?: Cluster['httpsProxy'];
-  noProxy?: Cluster['noProxy'];
-};
