--- src/common/types/trustedCertificateFields.ts
@@ -1,6 +1,0 @@
-import { InfraEnv } from '../api';
-
-export type TrustedCertificateFieldsType = {
-  enableCertificate: boolean;
-  trustBundle?: InfraEnv['additionalTrustBundle'];
-};
